✨ (core): Implement AnalysisService and AnalysisServiceImpl for financial overview data retrieval

+++ lib/presentation/bloc/scan_bill/scan_bill_state.dart
@@ -16,7 +16,7 @@ final class BillLoading extends ScanBillState {
 }
 
 class BillInfo {
-  final double money;
+  final int money;
   final DateTime date;
   final String store;
   final Category category;
