+++ lib/utils/currency_formatter.dart
@@ -1,7 +1,7 @@
 import 'package:intl/intl.dart';
 
 class CurrencyFormatter {
-  static String formatCurrency(double amount) {
+  static String formatCurrency(int amount) {
     return NumberFormat.simpleCurrency(locale: 'vi').format(amount);
   }
 }
